Fix generated unit test for Rails component gen
BUG: The generated unit tests for component generators were unable to
run the generator in the test because the Rails generator assumed a
Rails environment was loaded, which was not the case - see
<http://gist.github.com/95762>.

I have changed this so that by default the setup method in a
Rails-scoped component generators now generates a full Rails environment
in the temporary directory of the test (in the same way as the `rails`
command does). The directory is deleted again in the teardown method.

TODO: Silencing the generator output currently isn't working, so running
the tests produces a lot of unwanted console noise.